Starting to move each panel to a subdirectory
